asset components reorganization
