+++ src/components/asset/New/CloudStorage/index.test.js
@@ -1,12 +1,12 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 
-import CloudStorageActions from './CloudStorageActions'
+import CloudStorage from '.'
 
-it('CloudStorageActions renders without crashing', () => {
+it('CloudStorage renders without crashing', () => {
     const div = document.createElement('div')
     ReactDOM.render(
-        <CloudStorageActions linkSetter={() => null} />,
+        <CloudStorage linkSetter={() => null} />,
         div
     )
     ReactDOM.unmountComponentAtNode(div)
