+++ src/containers/CloudStorageLoader.js
@@ -1,12 +1,6 @@
 import { connect } from 'react-redux'
-import CloudStoragePicker from '../components/asset/CloudStoragePicker'
+import CloudStorage from '../components/asset/New/CloudStorage'
 
 export default connect(state => ({
-    blobs: Object.values(state.cloudStorage.blobs).map((fileObject, index) => ({
-        id: index,
-        container: fileObject.container,
-        blobName: fileObject.blobName,
-        value: false
-    })),
-    error: state.cloudStorage.error
-}))(CloudStoragePicker)
+    oauthAccounts: state.oauthAccounts
+}))(CloudStorage)
